feature/big-refactor-1 	1. Moved functions logging to component from page; 	2. Moved functions registering to component from page;
