+++ react/src/routes/LoginPage.jsx
@@ -1,0 +1,12 @@
+import React, {useState} from 'react';
+import Login from "../components/Login.jsx";
+
+const LoginPage = () => {
+    return (
+        <div className='wrapper'>
+            <Login/>
+        </div>
+    );
+};
+
+export default LoginPage;
