+++ react/src/routes/RegisteringPage.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import Registering from "../components/Registering.jsx";
+
+
+const RegisteringPage = () => {
+
+    return (
+        <div className='wrapper'>
+            <Registering/>
+        </div>
+    );
+};
+
+export default RegisteringPage;
