+++ react/src/routes/UsersPage.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const UsersPage = () => {
+    return (
+        <div>
+            Пользователи
+        </div>
+    );
+};
+
+export default UsersPage;
